requirement.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.7.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==4.2.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.2
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.40.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
